SUP-959 #comment fix non matching value types error

+++ generator/sources/java/src/com/kaltura/client/KalturaParamsValueDefaults.java
@@ -39,8 +39,8 @@ public final class KalturaParamsValueDefaults {
 	public static final float KALTURA_UNDEF_FLOAT = Float.MIN_VALUE;
 	public static final float KALTURA_NULL_FLOAT =  Float.MAX_VALUE;
 	
-	public static final int KALTURA_UNDEF_LONG =  Long.MIN_VALUE;
-	public static final int KALTURA_NULL_LONG =  Long.MAX_VALUE;
+	public static final long KALTURA_UNDEF_LONG =  Long.MIN_VALUE;
+	public static final long KALTURA_NULL_LONG =  Long.MAX_VALUE;
 	
 	public static final String KALTURA_NULL_STRING =  "__null_string__";
